refactor: add gc_stringview_arrays util and use that
